code/requirements.txt
astroid==2.5.6
azure-common==1.1.27
azure-core==1.13.0
azure-servicebus==7.1.1
azure-storage-queue==12.1.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
msrest==0.6.21
oauthlib==3.1.0
pycparser==2.20
pylint==2.8.2
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.19.0
six==1.15.0
toml==0.10.2
uamqp==1.3.0
urllib3==1.26.4
wrapt==1.12.1
